requirements.txt
Mezzanine
django-rest-framework
scikit-learn
numpy
scipy
